Split main.go into per-command files
